chore(samples): renamed samples to be used as shared files in docs
