--- samples/grids/tree-grid/keyboard-custom-navigation-tgrid/ReadMe.md
@@ -1,56 +1,0 @@
-<!-- NOTE: do not change this file because it's auto re-generated from template: -->
-<!-- https://github.com/IgniteUI/igniteui-react-examples/tree/vnext/templates/sample/ReadMe.md -->
-
-This folder contains implementation of React application with example of Keyboard Custom Navigation feature using [Grid](https://www.infragistics.com/products/ignite-ui-react/react/components/general-getting-started.html) component.
-
-
-<html lang="en" xmlns="http://www.w3.org/1999/xhtml">
-    <body>
-        <a target="_blank" href="https://www.infragistics.com/products/ignite-ui-react/react/components/general-getting-started.html" rel="noopener noreferrer">
-            <img height="40px" style="border-radius: 0rem" alt="View Docs" src="https://github.com/IgniteUI/igniteui-blazor-examples/raw/vnext/templates/sample/images/button-docs.png"/>
-        </a>
-        <a target="_blank" href="./src/index.tsx" rel="noopener noreferrer">
-            <img height="40px" style="border-radius: 0rem; max-width: 100%;" alt="View Code" src="https://github.com/IgniteUI/igniteui-blazor-examples/raw/vnext/templates/sample/images/button-code.png"/>
-        </a>
-        <a target="_blank" href="https://www.infragistics.com/react-demos/samples/grids/grid/keyboard-custom-navigation" rel="noopener noreferrer">
-            <img height="40px" style="border-radius: 0rem; max-width: 100%;" alt="Run Sample" src="https://github.com/IgniteUI/igniteui-blazor-examples/raw/vnext/templates/sample/images/button-run.png"/>
-        </a>
-        <a target="_blank" href="https://codesandbox.io/s/github/IgniteUI/igniteui-react-examples/tree/master/samples/grids/grid/keyboard-custom-navigation?fontsize=14&hidenavigation=1&theme=dark&view=preview&file=/src/index.tsx" rel="noopener noreferrer">
-            <img height="40px" style="border-radius: 0rem; max-width: 100%;" alt="Run Sample" src="https://github.com/IgniteUI/igniteui-blazor-examples/raw/vnext/templates/sample/images/button-sandbox.png"/>
-        </a>
-    </body>
-</html>
-
-## Branches
-
-> **_NOTE:_** You should use [master](https://github.com/IgniteUI/igniteui-react-examples/tree/master) branch of this repository if you want to run samples on your computer. Use the [vnext](https://github.com/IgniteUI/igniteui-react-examples/tree/vnext) branch only when you want to contribute new samples to this repository.
-
-## Instructions
-
-Follow these instructions to run this example:
-
-
-```
-git clone https://github.com/IgniteUI/igniteui-react-examples.git
-git checkout master
-cd ./igniteui-react-examples
-cd ./samples/grids/grid/keyboard-custom-navigation
-```
-
-open above folder in VS Code or type:
-```
-code .
-```
-
-In terminal window, run:
-```
-npm install --legacy-peer-deps
-npm run-script start
-```
-
-Then open http://localhost:4200/ in your browser
-
-
-## Learn More
-
-To learn more about **Ignite UI for React** components, check out the [React documentation](https://www.infragistics.com/products/ignite-ui-react/react/components/general-getting-started.html).
