--- samples/grids/tree-grid/keyboard-custom-navigation-tgrid/public/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <title>Sample | Ignite UI | React | infragistics</title>
-    <link rel="shortcut icon" href="https://static.infragistics.com/xplatform/images/browsers/react.png" >
-    <link rel="stylesheet"    href="https://static.infragistics.com/xplatform/css/samples/shared.v8.css" />
-  </head>
-  <body>
-    <div id="root"></div>
-  </body>
-</html>
