--- samples/grids/tree-grid/keyboard-custom-navigation-tgrid/sandbox.config.json
@@ -1,5 +1,0 @@
-{
-    "infiniteLoopProtection": false,
-    "hardReloadOnChange": false,
-    "view": "browser"
-}
