--- samples/grids/tree-grid/keyboard-custom-navigation-tgrid/src/index.css
@@ -1,7 +1,0 @@
-/* shared styles are loaded from: */
-/* https://static.infragistics.com/xplatform/css/samples */
-
-    #grid {
-        --ig-size: var(--ig-size-medium);
-    }
-
