--- samples/grids/tree-grid/keyboard-custom-navigation-tgrid/src/react-app-env.d.ts
@@ -1,1 +1,0 @@
-/// <reference types="react-scripts" />
