+++ samples/grids/tree-grid/keyboard-custom-navigation/src/index.css
@@ -1,2 +1,7 @@
 /* shared styles are loaded from: */
 /* https://static.infragistics.com/xplatform/css/samples */
+
+    #grid {
+        --ig-size: var(--ig-size-medium);
+    }
+
